+++ src/ninja/mbedded/ninjaterm/view/mainWindow/statusBar/StatusBarViewController.java
@@ -1,19 +1,20 @@
-package ninja.mbedded.ninjaterm.view.mainWindow.StatusBar;
+package ninja.mbedded.ninjaterm.view.mainWindow.statusBar;
 
 import javafx.beans.property.SimpleDoubleProperty;
+import javafx.beans.property.SimpleIntegerProperty;
 import javafx.beans.value.ChangeListener;
 import javafx.collections.ListChangeListener;
 import javafx.fxml.FXML;
-import javafx.fxml.Initializable;
 import javafx.scene.Node;
 import javafx.scene.control.Label;
 import javafx.scene.control.ScrollPane;
 import javafx.scene.text.TextFlow;
 import ninja.mbedded.ninjaterm.model.Model;
+import ninja.mbedded.ninjaterm.model.terminal.Terminal;
+import ninja.mbedded.ninjaterm.model.terminal.txRx.RawDataReceivedListener;
 import ninja.mbedded.ninjaterm.view.led.Led;
 
-import java.net.URL;
-import java.util.ResourceBundle;
+import java.util.List;
 
 /**
  * Controller for the "Status Bar" which is part of the main window.
@@ -40,6 +41,16 @@ public class StatusBarViewController {
     @FXML
     public Led activityRxLed;
 
+    //==============================================//
+    //================ BOTTOM BAR ==================//
+    //==============================================//
+
+    @FXML
+    public Label totalByteCountTx;
+
+    @FXML
+    public Label totalByteCountRx;
+
     @FXML
     public Label totalBytesPerSecTx;
 
@@ -65,6 +76,29 @@ public class StatusBarViewController {
             // Auto-scroll the status scroll-pane to the last received status message
             statusScrollPane.setVvalue(statusTextFlow.getHeight());
         });
+
+        //==============================================//
+        //============ TOTAL BYTE COUNT SETUP ==========//
+        //==============================================//
+
+        //======================= TX ===================//
+        ChangeListener<Number> totalByteCountTxChangeListener = (observable, oldValue, newValue) -> {
+            totalByteCountTx.setText(Integer.toString(newValue.intValue()));
+        };
+        model.status.totalByteCountTx.addListener(totalByteCountTxChangeListener);
+
+        // Set default (giving bogus data as it is not used)
+        totalByteCountTxChangeListener.changed(new SimpleIntegerProperty(), 0.0, 0.0);
+
+        //======================= RX ===================//
+        ChangeListener<Number> totalByteCountRxChangeListener = (observable, oldValue, newValue) -> {
+            totalByteCountRx.setText(Integer.toString(newValue.intValue()));
+        };
+        model.status.totalByteCountRx.addListener(totalByteCountRxChangeListener);
+
+        // Set default (giving bogus data as it is not used)
+        totalByteCountRxChangeListener.changed(new SimpleIntegerProperty(), 0.0, 0.0);
+
 
         //==============================================//
         //=========== TOTAL BYTES/SECOND SETUP =========//